+++ android-sdk/src/test/java/co/elastic/otel/android/functional/AttributesTest.kt
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package co.elastic.otel.android.integration
+package co.elastic.otel.android.functional
 
 import android.Manifest
 import android.content.Context
